refactor: slim down page and document
